[Guardian] Organize cryptographic type modules
