--- crates/hashi-types/src/guardian/crypto.rs
@@ -1,1223 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-use super::errors::GuardianError::InvalidInputs;
-use super::errors::GuardianResult;
-use super::kp_certs_roster::KPFingerprint;
-use super::kp_certs_roster::KpCerts;
-use super::kp_certs_roster::KpCertsRoster;
-use crate::bitcoin::BTC_LIB;
-use crate::pgp::PgpPublicCert;
-use crate::pgp::cert_owns_key_handle;
-use crate::pgp::encrypt_armored;
-use crate::pgp::pgp_message_recipients;
-use blake2::Blake2b;
-use blake2::Digest;
-use blake2::digest::consts::U32;
-use hpke::Deserializable;
-use hpke::Kem;
-use hpke::Serializable;
-use hpke::aead::AesGcm256;
-use hpke::kdf::HkdfSha384;
-use hpke::kem::X25519HkdfSha256;
-use k256::CompressedPoint;
-use k256::ProjectivePoint;
-use k256::Scalar;
-use k256::Secp256k1;
-use k256::elliptic_curve::Field;
-use k256::elliptic_curve::ScalarPrimitive;
-use k256::elliptic_curve::group::GroupEncoding;
-use rand_core::CryptoRng;
-use rand_core::RngCore;
-use serde::Deserialize;
-use serde::Serialize;
-use std::collections::BTreeMap;
-use std::collections::HashSet;
-use std::fmt;
-use std::num::NonZeroU16;
-use tracing::info;
-// ---------------------------------
-//      Crypto Structs & Types
-// ---------------------------------
-
-pub type EncSecKey = <X25519HkdfSha256 as Kem>::PrivateKey;
-pub type EncPubKey = <X25519HkdfSha256 as Kem>::PublicKey;
-pub type EncPubKeyBytes = Vec<u8>; // Use as an alternative to EncPubKey where Serialize is needed
-pub struct GuardianEncKeyPair {
-    sk: EncSecKey,
-    pk: EncPubKey,
-}
-pub type EncapsulatedKey = <X25519HkdfSha256 as Kem>::EncappedKey;
-
-pub type ShareID = NonZeroU16; // Share IDs are assigned from 1, e.g., 1, 2, 3 and so on.
-
-#[derive(Copy, Clone)]
-pub struct Share {
-    pub id: ShareID,
-    pub value: Scalar,
-}
-
-/// Minimum reconstruction threshold (`t > 1`).
-pub const MIN_THRESHOLD: usize = 2;
-/// Maximum total number of shares (`n <= u16::MAX`)
-pub const MAX_NUM_SHARES: usize = u16::MAX as usize;
-
-/// Validated `(n, t)` secret-sharing parameters.
-#[derive(Serialize, Deserialize, Copy, Clone, Debug, PartialEq, Eq)]
-pub struct SecretSharingParams {
-    num_shares: usize,
-    threshold: usize,
-}
-
-impl SecretSharingParams {
-    pub fn new(num_shares: usize, threshold: usize) -> GuardianResult<Self> {
-        if threshold < MIN_THRESHOLD {
-            return Err(InvalidInputs(format!(
-                "threshold {threshold} below minimum {MIN_THRESHOLD}"
-            )));
-        }
-        if num_shares < threshold {
-            return Err(InvalidInputs(format!(
-                "num_shares {num_shares} below threshold {threshold}"
-            )));
-        }
-        if num_shares > MAX_NUM_SHARES {
-            return Err(InvalidInputs(format!(
-                "{num_shares} must be at most u16::MAX"
-            )));
-        }
-        Ok(Self {
-            num_shares,
-            threshold,
-        })
-    }
-
-    pub fn num_shares(&self) -> usize {
-        self.num_shares
-    }
-
-    pub fn threshold(&self) -> usize {
-        self.threshold
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct GuardianEncryptedShare {
-    pub id: ShareID,
-    pub ciphertext: Ciphertext,
-}
-
-/// The encrypted copies of one secret share assigned to one KP, with one
-/// ciphertext per accepted certificate.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct KPEncryptedShares {
-    pub id: ShareID,
-    /// Recipient PGP fingerprint to armored ciphertext.
-    pub ciphertexts_by_fingerprint: BTreeMap<KPFingerprint, String>,
-}
-
-/// The complete encrypted-share roster for a ceremony, canonicalized by share
-/// id. Each entry contains one KP's ciphertexts.
-#[derive(Serialize, Debug, Clone, PartialEq)]
-pub struct KPEncryptedSharesRoster(Vec<KPEncryptedShares>);
-
-pub type DigestBytes = Vec<u8>;
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct ShareCommitment {
-    pub id: ShareID,
-    pub digest: DigestBytes,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct ShareCommitments(
-    #[serde(with = "crate::guardian::serde::hex_map_values")] BTreeMap<ShareID, DigestBytes>,
-);
-
-/// Public description of the current BTC key's secret-sharing scheme.
-/// `sharing_seq` versions the instance: setup writes 0, each rotation bumps it by 1.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct SecretSharingInstance {
-    commitments: ShareCommitments,
-    params: SecretSharingParams,
-    sharing_seq: u64,
-}
-
-impl SecretSharingInstance {
-    pub fn new(
-        commitments: ShareCommitments,
-        num_shares: usize,
-        threshold: usize,
-        sharing_seq: u64,
-    ) -> GuardianResult<Self> {
-        let params = SecretSharingParams::new(num_shares, threshold)?;
-        if commitments.len() != params.num_shares() {
-            return Err(InvalidInputs(format!(
-                "expected {} commitments, got {}",
-                params.num_shares(),
-                commitments.len()
-            )));
-        }
-        let commitment_ids: Vec<u16> = commitments.iter().map(|c| c.id.get()).collect();
-        let expected_ids: Vec<u16> = (1..=params.num_shares() as u16).collect();
-        if commitment_ids != expected_ids {
-            return Err(InvalidInputs(format!(
-                "commitment ids are not exactly 1..={}: got {commitment_ids:?}",
-                params.num_shares()
-            )));
-        }
-        Ok(Self {
-            commitments,
-            params,
-            sharing_seq,
-        })
-    }
-
-    pub fn commitments(&self) -> &ShareCommitments {
-        &self.commitments
-    }
-
-    pub fn num_shares(&self) -> usize {
-        self.params.num_shares()
-    }
-
-    pub fn threshold(&self) -> usize {
-        self.params.threshold()
-    }
-
-    pub fn sharing_seq(&self) -> u64 {
-        self.sharing_seq
-    }
-}
-
-impl fmt::Display for SecretSharingInstance {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let bytes = bcs::to_bytes(&self.commitments).expect("serialization should work");
-        let commitments_hash = Blake2b::<U32>::digest(bytes);
-        let commitments_hash_hex = hex::encode(commitments_hash);
-        write!(
-            f,
-            "SecretSharingInstance(seq={},n={},t={},commitments_hash={})",
-            self.sharing_seq,
-            self.num_shares(),
-            self.threshold(),
-            commitments_hash_hex
-        )
-    }
-}
-
-impl ShareCommitments {
-    pub fn new(commitments: Vec<ShareCommitment>) -> GuardianResult<Self> {
-        let mut map = BTreeMap::new();
-        for commitment in commitments {
-            if map.insert(commitment.id, commitment.digest).is_some() {
-                return Err(InvalidInputs("duplicate share id".into()));
-            }
-        }
-        Ok(Self(map))
-    }
-
-    pub fn from_shares(shares: &[Share]) -> GuardianResult<Self> {
-        Self::new(shares.iter().map(commit_share).collect())
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    pub fn contains(&self, commitment: &ShareCommitment) -> bool {
-        self.0
-            .get(&commitment.id)
-            .is_some_and(|digest| digest == &commitment.digest)
-    }
-
-    /// Verify `share`'s commitment is in this set.
-    pub fn verify_share(&self, share: &Share) -> GuardianResult<()> {
-        self.contains(&commit_share(share))
-            .then_some(())
-            .ok_or_else(|| InvalidInputs("No matching share found".into()))
-    }
-
-    pub fn iter(&self) -> impl Iterator<Item = ShareCommitment> + '_ {
-        self.0.iter().map(|(id, digest)| ShareCommitment {
-            id: *id,
-            digest: digest.clone(),
-        })
-    }
-}
-
-impl IntoIterator for ShareCommitments {
-    type Item = ShareCommitment;
-    type IntoIter = std::vec::IntoIter<ShareCommitment>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.0
-            .into_iter()
-            .map(|(id, digest)| ShareCommitment { id, digest })
-            .collect::<Vec<_>>()
-            .into_iter()
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct Ciphertext {
-    pub encapsulated_key: Vec<u8>,
-    pub aes_ciphertext: Vec<u8>,
-}
-
-// ---------------------------------
-//          Helper impl's
-// ---------------------------------
-
-impl GuardianEncKeyPair {
-    pub fn random<R: CryptoRng + RngCore>(rng: &mut R) -> Self {
-        let (sk, pk) = X25519HkdfSha256::gen_keypair(rng);
-        Self { sk, pk }
-    }
-
-    pub fn secret_key(&self) -> &EncSecKey {
-        &self.sk
-    }
-
-    pub fn public_key(&self) -> &EncPubKey {
-        &self.pk
-    }
-}
-
-impl KPEncryptedShares {
-    /// Verify that this share has exactly one ciphertext per expected cert and
-    /// that every ciphertext targets only keys owned by its keyed cert.
-    pub fn verify_recipients(&self, certs: &KpCerts) -> GuardianResult<()> {
-        let expected_fingerprints = certs.fingerprints();
-        let actual_fingerprints = self
-            .ciphertexts_by_fingerprint
-            .keys()
-            .cloned()
-            .collect::<Vec<_>>();
-        if actual_fingerprints != expected_fingerprints {
-            return Err(InvalidInputs(format!(
-                "encrypted share recipient roster differs for share id {}: expected {:?}, got \
-                 {:?}",
-                self.id.get(),
-                expected_fingerprints,
-                actual_fingerprints
-            )));
-        }
-
-        for cert in certs.pgp_certs() {
-            let fingerprint = cert.fingerprint().to_hex();
-            let ciphertext = self
-                .ciphertexts_by_fingerprint
-                .get(&fingerprint)
-                .expect("fingerprint rosters were checked for equality");
-            verify_pgp_ciphertext_recipient(self.id, &fingerprint, ciphertext, cert)?;
-        }
-        Ok(())
-    }
-}
-
-impl KPEncryptedSharesRoster {
-    pub fn new(mut shares: Vec<KPEncryptedShares>) -> GuardianResult<Self> {
-        shares.sort_by_key(|s| s.id);
-
-        if shares.len() > MAX_NUM_SHARES {
-            return Err(InvalidInputs(format!(
-                "{} shares must be at most u16::MAX",
-                shares.len()
-            )));
-        }
-        let ids: Vec<u16> = shares.iter().map(|s| s.id.get()).collect();
-        let expected: Vec<u16> = (1..=shares.len() as u16).collect();
-        if ids != expected {
-            return Err(InvalidInputs(format!(
-                "encrypted share ids are not exactly 1..={}: got {ids:?}",
-                shares.len()
-            )));
-        }
-
-        let mut seen_fingerprints = HashSet::new();
-        for share in &shares {
-            if share.ciphertexts_by_fingerprint.is_empty() {
-                return Err(InvalidInputs(format!(
-                    "encrypted share id {} must have at least one PGP ciphertext",
-                    share.id.get()
-                )));
-            }
-            for fingerprint in share.ciphertexts_by_fingerprint.keys() {
-                if !seen_fingerprints.insert(fingerprint.clone()) {
-                    return Err(InvalidInputs(format!(
-                        "duplicate encrypted share recipient fingerprint {}",
-                        fingerprint
-                    )));
-                }
-            }
-        }
-
-        Ok(Self(shares))
-    }
-
-    /// Number of secret shares, not the number of cert-wrapped
-    /// ciphertexts. A KP may have multiple ciphertexts for the same share id.
-    pub fn share_count(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn ciphertext_count(&self) -> usize {
-        self.0
-            .iter()
-            .map(|share| share.ciphertexts_by_fingerprint.len())
-            .sum()
-    }
-
-    pub fn iter(&self) -> impl Iterator<Item = &KPEncryptedShares> {
-        self.0.iter()
-    }
-
-    pub fn into_vec(self) -> Vec<KPEncryptedShares> {
-        self.0
-    }
-
-    pub fn find_by_fingerprint(&self, fingerprint: &str) -> Option<(&KPEncryptedShares, &str)> {
-        self.iter().find_map(|share| {
-            share
-                .ciphertexts_by_fingerprint
-                .get(fingerprint)
-                .map(|ciphertext| (share, ciphertext.as_str()))
-        })
-    }
-
-    /// Require `fingerprint` to be assigned to `submitted_share_id` in this
-    /// encrypted-share roster.
-    pub fn validate_share_assignment(
-        &self,
-        fingerprint: &str,
-        submitted_share_id: ShareID,
-    ) -> GuardianResult<()> {
-        let (assigned_share, _) = self.find_by_fingerprint(fingerprint).ok_or_else(|| {
-            InvalidInputs(format!(
-                "KP fingerprint {fingerprint} is not present in the encrypted-share roster"
-            ))
-        })?;
-        if assigned_share.id != submitted_share_id {
-            return Err(InvalidInputs(format!(
-                "KP fingerprint {fingerprint} is assigned share id {}, not submitted share id {}",
-                assigned_share.id.get(),
-                submitted_share_id.get()
-            )));
-        }
-        Ok(())
-    }
-
-    /// Replace one certificate-specific ciphertext while preserving every
-    /// other KP/share entry and revalidating global fingerprint uniqueness.
-    pub fn replace_recipient(
-        self,
-        current_fingerprint: &str,
-        new_fingerprint: KPFingerprint,
-        new_ciphertext: String,
-    ) -> GuardianResult<(Self, KPEncryptedShares)> {
-        if new_fingerprint == current_fingerprint {
-            return Err(InvalidInputs(format!(
-                "replacement KP fingerprint {new_fingerprint} must differ from the current \
-                 fingerprint"
-            )));
-        }
-        if self.find_by_fingerprint(&new_fingerprint).is_some() {
-            return Err(InvalidInputs(format!(
-                "new KP fingerprint {new_fingerprint} is already present in the encrypted share \
-                 roster"
-            )));
-        }
-
-        let mut shares = self.0;
-        let share = shares
-            .iter_mut()
-            .find(|share| {
-                share
-                    .ciphertexts_by_fingerprint
-                    .contains_key(current_fingerprint)
-            })
-            .ok_or_else(|| {
-                InvalidInputs(format!(
-                    "current KP fingerprint {current_fingerprint} is not present in the \
-                     encrypted share roster"
-                ))
-            })?;
-        share
-            .ciphertexts_by_fingerprint
-            .remove(current_fingerprint)
-            .expect("the current fingerprint was located in this entry");
-        share
-            .ciphertexts_by_fingerprint
-            .insert(new_fingerprint, new_ciphertext);
-        let changed_share = share.clone();
-        Ok((Self::new(shares)?, changed_share))
-    }
-
-    /// Recipient PGP fingerprints grouped by share id.
-    pub fn recipient_roster(&self) -> Vec<Vec<KPFingerprint>> {
-        let mut grouped: Vec<Vec<KPFingerprint>> = Vec::with_capacity(self.share_count());
-        for share in &self.0 {
-            grouped.push(share.ciphertexts_by_fingerprint.keys().cloned().collect());
-        }
-        grouped
-    }
-
-    /// Verify every encrypted share against the certs assigned to its share id.
-    pub fn verify_recipients(&self, certs_roster: &KpCertsRoster) -> GuardianResult<()> {
-        if self.share_count() != certs_roster.num_kps() {
-            return Err(InvalidInputs(format!(
-                "expected {} KP cert roster entries, got {} encrypted shares",
-                certs_roster.num_kps(),
-                self.share_count()
-            )));
-        }
-
-        for share in self.iter() {
-            let certs = certs_roster
-                .certs_for_share(share.id)
-                .expect("encrypted share ids are exactly 1..=roster length");
-            share.verify_recipients(certs)?;
-        }
-        Ok(())
-    }
-}
-
-impl<'de> Deserialize<'de> for KPEncryptedSharesRoster {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let shares = Vec::<KPEncryptedShares>::deserialize(deserializer)?;
-        Self::new(shares).map_err(serde::de::Error::custom)
-    }
-}
-
-fn verify_pgp_ciphertext_recipient(
-    share_id: ShareID,
-    recipient_fingerprint: &str,
-    ciphertext: &str,
-    expected_cert: &PgpPublicCert,
-) -> GuardianResult<()> {
-    let recipients = pgp_message_recipients(ciphertext).map_err(|e| {
-        InvalidInputs(format!(
-            "failed to parse PGP recipients for share id {}: {e}",
-            share_id.get()
-        ))
-    })?;
-    if recipients.is_empty() {
-        return Err(InvalidInputs(format!(
-            "share id {} has no PGP recipients",
-            share_id.get()
-        )));
-    }
-    for handle in &recipients {
-        if !cert_owns_key_handle(expected_cert, handle) {
-            return Err(InvalidInputs(format!(
-                "share id {} (keyed by {}) is encrypted to key {handle}, which is not in that \
-                 cert",
-                share_id.get(),
-                recipient_fingerprint
-            )));
-        }
-    }
-    info!(
-        share_id = share_id.get(),
-        fingerprint = %recipient_fingerprint,
-        recipient_count = recipients.len(),
-        "verified encrypted share targets only its keyed recipient cert"
-    );
-    Ok(())
-}
-
-pub fn to_scalar(id: ShareID) -> Scalar {
-    Scalar::from(id.get() as u32)
-}
-
-// ---------------------------------
-//    Encryption/Decryption utils
-// ---------------------------------
-
-/// Encrypts plaintext. Returns InvalidInputs if plaintext / aad is extraordinarily long (~2^36).
-pub fn encrypt<R: CryptoRng + RngCore>(
-    bytes: &[u8],
-    pk: &EncPubKey,
-    aad: Option<&[u8; 32]>,
-    rng: &mut R,
-) -> GuardianResult<Ciphertext> {
-    let (encapsulated_key, aes_ciphertext) =
-        hpke::single_shot_seal::<AesGcm256, HkdfSha384, X25519HkdfSha256, _>(
-            &hpke::OpModeS::Base,
-            pk,
-            &[],
-            bytes,
-            aad.unwrap_or(&[0; 32]),
-            rng,
-        )
-        .map_err(|e| InvalidInputs(format!("Encryption failed: {}", e)))?;
-    Ok(Ciphertext {
-        encapsulated_key: encapsulated_key.to_bytes().to_vec(),
-        aes_ciphertext,
-    })
-}
-
-/// Decrypts ciphertext. Returns InvalidInputs if aad is invalid.
-pub fn decrypt(
-    ciphertext: &Ciphertext,
-    sk: &EncSecKey,
-    aad: Option<&[u8; 32]>,
-) -> GuardianResult<Vec<u8>> {
-    let encapsulated_key = EncapsulatedKey::from_bytes(&ciphertext.encapsulated_key)
-        .map_err(|e| InvalidInputs(format!("Failed to deserialize encapsulated key: {}", e)))?;
-    hpke::single_shot_open::<AesGcm256, HkdfSha384, X25519HkdfSha256>(
-        &hpke::OpModeR::Base,
-        sk,
-        &encapsulated_key,
-        &[],
-        &ciphertext.aes_ciphertext,
-        aad.unwrap_or(&[0; 32]),
-    )
-    .map_err(|e| InvalidInputs(format!("Decryption failed: {}", e)))
-}
-
-// ---------------------------------
-//    Secret-sharing utilities
-// ---------------------------------
-
-/// Split a k256 SecretKey into `params.num_shares()` shares using Shamir's
-/// secret-sharing with reconstruction threshold `params.threshold()`.
-pub fn split_secret<R: CryptoRng + RngCore>(
-    sk: &k256::SecretKey,
-    params: &SecretSharingParams,
-    rng: &mut R,
-) -> Vec<Share> {
-    let secret = *sk.to_nonzero_scalar().as_ref();
-    let mut coefficients = vec![secret];
-    for _ in 0..(params.threshold() - 1) {
-        coefficients.push(Scalar::random(&mut *rng))
-    }
-
-    (1..=params.num_shares())
-        .map(|i| NonZeroU16::new(i as u16).expect("validated num_shares fits in u16"))
-        .map(|i| Share {
-            id: i,
-            value: eval_poly(i, &coefficients),
-        })
-        .collect()
-}
-
-// Coefficients: [c0, c1, c2, c3]
-// Returns: c0 + c1 * x + c2 * x^2 + c3 * x^3
-pub fn eval_poly(pos: ShareID, coefficients: &[Scalar]) -> Scalar {
-    let x = to_scalar(pos);
-    let mut out = Scalar::ZERO;
-    let mut xpow = Scalar::ONE;
-    for c in coefficients {
-        out = out.add(&c.mul(&xpow));
-        xpow = xpow.mul(&x);
-    }
-    out
-}
-
-/// Combine secret shares into the reconstructed `k256::SecretKey` with
-/// threshold `t`. Errors on duplicate share IDs or fewer than `t` shares.
-/// Callers that need a `bitcoin::secp256k1::Keypair` should pass the result
-/// through `k256_sk_to_btc_keypair`.
-pub fn combine_shares(shares: &[Share], t: usize) -> GuardianResult<k256::SecretKey> {
-    // Validation: ensure no duplicates
-    let mut seen_ids = std::collections::HashSet::new();
-    for share in shares {
-        if !seen_ids.insert(share.id) {
-            return Err(InvalidInputs("Duplicate share ID".into()));
-        }
-    }
-    if seen_ids.len() < t {
-        return Err(InvalidInputs(format!(
-            "Received only {} out of {} shares",
-            seen_ids.len(),
-            t
-        )));
-    }
-
-    let ids = shares.iter().map(|s| to_scalar(s.id)).collect::<Vec<_>>();
-    let mut result = Scalar::ZERO;
-    for share in shares {
-        let cur_share_id = to_scalar(share.id);
-        let numerator: Scalar = ids
-            .iter()
-            .filter(|&id| cur_share_id != *id)
-            .map(|id| id.negate())
-            .product();
-        let denominator: Scalar = ids
-            .iter()
-            .filter(|&id| cur_share_id != *id)
-            .map(|id| cur_share_id.sub(id))
-            .product();
-
-        // Lagrange basis polynomial evaluated at x=0
-        // L_i(0) = product_{j != i} (-x_j) / (x_i - x_j)
-        let lagrange_basis = numerator.mul(
-            &denominator
-                .invert()
-                .expect("Denominator is never zero because share IDs are unique"),
-        );
-        result = result.add(&share.value.mul(&lagrange_basis));
-    }
-
-    info!("Bitcoin key created with fingerprint {:x}", exp_g(&result));
-
-    Ok(k256::SecretKey::from_slice(&result.to_bytes())
-        .expect("k256 scalar bytes are a valid k256 secret key"))
-}
-
-/// Convert a `k256::SecretKey` to a `bitcoin::secp256k1::Keypair`. Both libs
-/// use big-endian 32-byte scalars so the byte round-trip is value-preserving.
-/// We juggle between the two libraries because secp256k1 does not expose the
-/// arithmetic tools needed for secret-sharing.
-pub fn k256_sk_to_btc_keypair(sk: &k256::SecretKey) -> bitcoin::secp256k1::Keypair {
-    let btc_sk = bitcoin::secp256k1::SecretKey::from_slice(&sk.to_bytes())
-        .expect("k256 secret key bytes are a valid secp256k1 secret key");
-    bitcoin::secp256k1::Keypair::from_secret_key(&BTC_LIB, &btc_sk)
-}
-
-/// The x-only BTC public key for a `k256::SecretKey` — the guardian's BTC master
-/// pubkey. Bridges through the secp256k1 keypair, like [`k256_sk_to_btc_keypair`].
-pub fn k256_sk_to_btc_xonly_pubkey(sk: &k256::SecretKey) -> crate::bitcoin::BitcoinPubkey {
-    k256_sk_to_btc_keypair(sk).x_only_public_key().0
-}
-
-/// Create a commitment (hash) for a share
-pub fn commit_share(share: &Share) -> ShareCommitment {
-    let commitment = ProjectivePoint::GENERATOR * share.value;
-    ShareCommitment {
-        id: share.id,
-        digest: commitment.to_bytes().to_vec(),
-    }
-}
-
-/// Encrypt a share with optional AAD
-pub fn encrypt_share<R: CryptoRng + RngCore>(
-    share: &Share,
-    pk: &EncPubKey,
-    aad: Option<&[u8; 32]>,
-    rng: &mut R,
-) -> GuardianEncryptedShare {
-    GuardianEncryptedShare {
-        id: share.id,
-        ciphertext: encrypt(&share.value.to_bytes(), pk, aad, rng)
-            .expect("neither plaintext nor aad are long"),
-    }
-}
-
-/// Split `sk` into `params.num_shares()` shares with reconstruction threshold
-/// `params.threshold()`, encrypt each share to every cert in the matching KP
-/// cert collection, and compute one commitment per share. The roster assigns
-/// its entry at position `i` to share ID `i + 1`.
-///
-/// # Panics
-///
-/// Panics if `kp_certs_roster.num_kps() != params.num_shares()`.
-pub fn split_and_encrypt_for_kps<R: CryptoRng + RngCore>(
-    sk: &k256::SecretKey,
-    kp_certs_roster: &KpCertsRoster,
-    params: &SecretSharingParams,
-    rng: &mut R,
-) -> (KPEncryptedSharesRoster, ShareCommitments) {
-    assert_eq!(
-        kp_certs_roster.num_kps(),
-        params.num_shares(),
-        "request validation ensures one KP cert collection per share",
-    );
-    let shares = split_secret(sk, params, rng);
-    let n = params.num_shares();
-    let mut encrypted_shares = Vec::with_capacity(n);
-    let mut commitments = Vec::with_capacity(n);
-    for (share, cert_set) in shares.iter().zip(kp_certs_roster.iter()) {
-        let ciphertexts_by_fingerprint = cert_set
-            .pgp_certs()
-            .iter()
-            .map(|cert| {
-                (
-                    cert.fingerprint().to_hex(),
-                    encrypt_share_for_provisioner(share, cert),
-                )
-            })
-            .collect();
-        encrypted_shares.push(KPEncryptedShares {
-            id: share.id,
-            ciphertexts_by_fingerprint,
-        });
-        commitments.push(commit_share(share));
-    }
-    let encrypted_shares = KPEncryptedSharesRoster::new(encrypted_shares)
-        .expect("split_secret produces share ids exactly 1..=n");
-    let commitments =
-        ShareCommitments::new(commitments).expect("share IDs 1..=n are unique by construction");
-    (encrypted_shares, commitments)
-}
-
-/// Encrypt a share for delivery to a key provisioner using OpenPGP ASCII armor.
-pub fn encrypt_share_for_provisioner(share: &Share, cert: &PgpPublicCert) -> String {
-    encrypt_armored(&share.value.to_bytes(), cert)
-        .expect("PgpPublicCert validation ensures OpenPGP encryption works")
-}
-
-/// Decrypt an encrypted share with optional AAD
-pub fn decrypt_share(
-    encrypted_share: &GuardianEncryptedShare,
-    sk: &EncSecKey,
-    aad: Option<&[u8; 32]>,
-) -> GuardianResult<Share> {
-    let serialized_share = decrypt(&encrypted_share.ciphertext, sk, aad)?;
-    let value = ScalarPrimitive::<Secp256k1>::from_slice(&serialized_share)
-        .map(Scalar::from)
-        .map_err(|_| InvalidInputs("Failed to deserialize share".into()))?;
-    Ok(Share {
-        id: encrypted_share.id,
-        value,
-    })
-}
-
-/// Decrypt each signed submission, verify it against the sharing instance, and
-/// reject duplicate share ids. The KP signature authenticates each ciphertext,
-/// so no additional HPKE AAD is needed.
-pub fn decrypt_verify_shares(
-    encrypted: &[GuardianEncryptedShare],
-    sk: &EncSecKey,
-    instance: &SecretSharingInstance,
-) -> GuardianResult<Vec<Share>> {
-    let share_count = encrypted.len();
-    let threshold = instance.threshold();
-    let num_shares = instance.num_shares();
-    if share_count < threshold {
-        return Err(InvalidInputs(format!(
-            "need at least {threshold} signed share submissions, got {share_count}"
-        )));
-    }
-    if share_count > num_shares {
-        return Err(InvalidInputs(format!(
-            "at most {num_shares} signed share submissions are allowed, got {share_count}"
-        )));
-    }
-
-    let mut shares: Vec<Share> = Vec::with_capacity(encrypted.len());
-    for enc in encrypted {
-        let share = decrypt_share(enc, sk, None)?;
-        instance.commitments().verify_share(&share)?;
-        if shares.iter().any(|s| s.id == share.id) {
-            return Err(InvalidInputs("Duplicate share ID".into()));
-        }
-        shares.push(share);
-    }
-    Ok(shares)
-}
-
-pub fn fingerprint(sk: &k256::SecretKey) -> CompressedPoint {
-    exp_g(&Scalar::from(sk.as_scalar_primitive()))
-}
-
-pub fn exp_g(scalar: &Scalar) -> CompressedPoint {
-    (ProjectivePoint::GENERATOR * scalar).to_bytes()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use k256::SecretKey;
-
-    #[test]
-    fn test_encrypt_and_decrypt() {
-        let bytes = b"Let's encrypt some stuff!";
-        let keypair = GuardianEncKeyPair::random(&mut rand::thread_rng());
-        let aad = Some(&[0; 32]);
-        let ciphertext =
-            encrypt(bytes, keypair.public_key(), aad, &mut rand::thread_rng()).unwrap();
-        assert!(decrypt(&ciphertext, keypair.secret_key(), aad).is_ok_and(|x| x == bytes));
-
-        let wrong_aad = Some(&[10; 32]);
-        assert!(
-            decrypt(&ciphertext, keypair.secret_key(), wrong_aad)
-                .is_err_and(|x| matches!(x, InvalidInputs(_)))
-        );
-    }
-
-    #[test]
-    fn decrypt_share_rejects_wrong_plaintext_length() {
-        let keypair = GuardianEncKeyPair::random(&mut rand::thread_rng());
-        for len in [31, 33] {
-            let encrypted_share = GuardianEncryptedShare {
-                id: ShareID::new(1).unwrap(),
-                ciphertext: encrypt(
-                    &vec![0; len],
-                    keypair.public_key(),
-                    None,
-                    &mut rand::thread_rng(),
-                )
-                .unwrap(),
-            };
-
-            let Err(err) = decrypt_share(&encrypted_share, keypair.secret_key(), None) else {
-                panic!("{len}-byte plaintext must be rejected");
-            };
-            assert!(matches!(err, InvalidInputs(_)), "{err:?}");
-        }
-    }
-
-    // Verify secret reconstruction with varying number of shares (0 to n).
-    // For each `num_shares`:
-    // - Below threshold: combine errors (refuses to interpolate)
-    // - Threshold or above: returns the original secret
-    fn check_reconstruction_with_varying_share_count(n: usize, t: usize) {
-        let original_k256_sk = SecretKey::random(&mut rand::thread_rng());
-        let original_bytes = original_k256_sk.to_bytes();
-        let shares = split_secret(
-            &original_k256_sk,
-            &SecretSharingParams::new(n, t).unwrap(),
-            &mut rand::thread_rng(),
-        );
-
-        for num_shares in 0..=n {
-            let result = combine_shares(&shares[0..num_shares], t);
-
-            if num_shares < t {
-                assert!(
-                    result.is_err(),
-                    "n={n} t={t} num_shares={num_shares}: subthreshold combine should error"
-                );
-            } else {
-                let reconstructed = result.unwrap();
-                assert_eq!(
-                    original_bytes,
-                    reconstructed.to_bytes(),
-                    "n={n} t={t} num_shares={num_shares}: should reconstruct original",
-                );
-            }
-        }
-    }
-
-    // Verify that certain other subsets of `t` shares reconstructs the original.
-    fn check_varying_subsets(n: usize, t: usize) {
-        let original_sk = SecretKey::random(&mut rand::thread_rng());
-        let original_bytes = original_sk.to_bytes();
-        let shares = split_secret(
-            &original_sk,
-            &SecretSharingParams::new(n, t).unwrap(),
-            &mut rand::thread_rng(),
-        );
-
-        for start_idx in 0..=(n - t) {
-            let subset = &shares[start_idx..(start_idx + t)];
-            let reconstructed = combine_shares(subset, t).unwrap();
-            assert_eq!(
-                original_bytes,
-                reconstructed.to_bytes(),
-                "n={n} t={t} start_idx={start_idx}: subset should reconstruct original",
-            );
-        }
-    }
-
-    fn check_combine_shares_rejects_duplicate_ids(n: usize, t: usize) {
-        let sk = SecretKey::random(&mut rand::thread_rng());
-        let shares = split_secret(
-            &sk,
-            &SecretSharingParams::new(n, t).unwrap(),
-            &mut rand::thread_rng(),
-        );
-
-        // First t-1 distinct shares plus a duplicate of shares[0].
-        let mut duplicate_shares: Vec<_> = shares.iter().take(t - 1).copied().collect();
-        duplicate_shares.push(shares[0]);
-
-        let err = combine_shares(&duplicate_shares, t)
-            .expect_err("combine_shares should reject duplicate share IDs");
-        assert!(
-            err.to_string().contains("Duplicate share ID"),
-            "n={n} t={t}: expected duplicate-id error, got {err}"
-        );
-    }
-
-    fn test_commitments(ids: &[u16]) -> ShareCommitments {
-        ShareCommitments::new(
-            ids.iter()
-                .map(|&id| ShareCommitment {
-                    id: NonZeroU16::new(id).unwrap(),
-                    digest: vec![id as u8],
-                })
-                .collect(),
-        )
-        .unwrap()
-    }
-
-    #[test]
-    fn share_commitments_json_encodes_digests_as_hex() {
-        let commitments = test_commitments(&[1, 2]);
-        let json = serde_json::to_value(&commitments).unwrap();
-
-        assert_eq!(json["1"], "01");
-        assert_eq!(json["2"], "02");
-        assert_eq!(
-            serde_json::from_value::<ShareCommitments>(json).unwrap(),
-            commitments
-        );
-    }
-
-    fn test_kp_encrypted_shares(id: u16) -> KPEncryptedShares {
-        test_kp_encrypted_shares_for_fingerprint(id, format!("fingerprint-{id}"))
-    }
-
-    fn test_kp_encrypted_shares_for_fingerprint(
-        id: u16,
-        recipient_fingerprint: String,
-    ) -> KPEncryptedShares {
-        test_kp_encrypted_shares_for_fingerprints(id, vec![recipient_fingerprint])
-    }
-
-    fn test_kp_encrypted_shares_for_fingerprints(
-        id: u16,
-        recipient_fingerprints: Vec<String>,
-    ) -> KPEncryptedShares {
-        KPEncryptedShares {
-            id: NonZeroU16::new(id).unwrap(),
-            ciphertexts_by_fingerprint: recipient_fingerprints
-                .into_iter()
-                .map(|recipient_fingerprint| {
-                    (
-                        recipient_fingerprint,
-                        "-----BEGIN PGP MESSAGE-----\n\n-----END PGP MESSAGE-----".into(),
-                    )
-                })
-                .collect(),
-        }
-    }
-
-    // Parameterized test cases: covers minimum (n=t=2), small, default, and large.
-    #[test]
-    fn reconstruction_with_varying_share_count_2_2() {
-        check_reconstruction_with_varying_share_count(2, 2);
-    }
-    #[test]
-    fn reconstruction_with_varying_share_count_3_2() {
-        check_reconstruction_with_varying_share_count(3, 2);
-    }
-    #[test]
-    fn reconstruction_with_varying_share_count_5_3() {
-        check_reconstruction_with_varying_share_count(5, 3);
-    }
-    #[test]
-    fn reconstruction_with_varying_share_count_10_7() {
-        check_reconstruction_with_varying_share_count(10, 7);
-    }
-
-    #[test]
-    fn varying_subsets_2_2() {
-        check_varying_subsets(2, 2);
-    }
-    #[test]
-    fn varying_subsets_3_2() {
-        check_varying_subsets(3, 2);
-    }
-    #[test]
-    fn varying_subsets_5_3() {
-        check_varying_subsets(5, 3);
-    }
-    #[test]
-    fn varying_subsets_10_7() {
-        check_varying_subsets(10, 7);
-    }
-
-    #[test]
-    fn combine_shares_rejects_duplicate_ids_2_2() {
-        check_combine_shares_rejects_duplicate_ids(2, 2);
-    }
-    #[test]
-    fn combine_shares_rejects_duplicate_ids_3_2() {
-        check_combine_shares_rejects_duplicate_ids(3, 2);
-    }
-    #[test]
-    fn combine_shares_rejects_duplicate_ids_5_3() {
-        check_combine_shares_rejects_duplicate_ids(5, 3);
-    }
-    #[test]
-    fn combine_shares_rejects_duplicate_ids_10_7() {
-        check_combine_shares_rejects_duplicate_ids(10, 7);
-    }
-
-    #[test]
-    fn secret_sharing_params_validation_cases() {
-        // Valid pairs.
-        for &(n, t) in &[(2, 2), (3, 2), (5, 3), (10, 7), (MAX_NUM_SHARES, 100)] {
-            SecretSharingParams::new(n, t)
-                .unwrap_or_else(|e| panic!("(n={n}, t={t}) should be valid: {e}"));
-        }
-        // Threshold below minimum.
-        for t in 0..MIN_THRESHOLD {
-            assert!(
-                SecretSharingParams::new(5, t).is_err(),
-                "t={t} (< MIN_THRESHOLD={MIN_THRESHOLD}) should be rejected"
-            );
-        }
-        // num_shares < threshold.
-        assert!(SecretSharingParams::new(2, 3).is_err());
-        assert!(SecretSharingParams::new(5, 7).is_err());
-        // num_shares > MAX_NUM_SHARES.
-        assert!(SecretSharingParams::new(MAX_NUM_SHARES + 1, 3).is_err());
-    }
-
-    #[test]
-    fn secret_sharing_instance_accepts_exact_commitment_ids() {
-        SecretSharingInstance::new(test_commitments(&[1, 2, 3]), 3, 2, 0)
-            .expect("commitment ids exactly 1..=n should be accepted");
-    }
-
-    #[test]
-    fn secret_sharing_instance_rejects_non_contiguous_commitment_ids() {
-        let err = SecretSharingInstance::new(test_commitments(&[1, 2, 4]), 3, 2, 0)
-            .expect_err("commitment ids must be exactly 1..=n");
-        assert!(format!("{err}").contains("commitment ids"), "{err}");
-    }
-
-    #[test]
-    fn secret_sharing_instance_display_is_concise() {
-        let instance = SecretSharingInstance::new(test_commitments(&[1, 2, 3]), 3, 2, 7)
-            .expect("valid instance");
-        let display = instance.to_string();
-
-        assert!(display.contains("seq=7"), "{display}");
-        assert!(display.contains("n=3"), "{display}");
-        assert!(display.contains("t=2"), "{display}");
-        assert!(display.contains("commitments_hash="), "{display}");
-    }
-
-    #[test]
-    fn kp_encrypted_shares_roster_canonicalizes_by_share_id_then_fingerprint() {
-        let shares = KPEncryptedSharesRoster::new(vec![
-            test_kp_encrypted_shares(3),
-            test_kp_encrypted_shares_for_fingerprints(
-                2,
-                vec!["fingerprint-2".into(), "fingerprint-2b".into()],
-            ),
-            test_kp_encrypted_shares(1),
-        ])
-        .unwrap();
-
-        let entries = shares
-            .iter()
-            .map(|s| {
-                (
-                    s.id.get(),
-                    s.ciphertexts_by_fingerprint
-                        .keys()
-                        .map(String::as_str)
-                        .collect::<Vec<_>>(),
-                )
-            })
-            .collect::<Vec<_>>();
-        assert_eq!(
-            entries,
-            vec![
-                (1, vec!["fingerprint-1"]),
-                (2, vec!["fingerprint-2", "fingerprint-2b"]),
-                (3, vec!["fingerprint-3"]),
-            ]
-        );
-        assert_eq!(shares.share_count(), 3);
-        assert_eq!(shares.ciphertext_count(), 4);
-        assert_eq!(
-            shares.recipient_roster(),
-            vec![
-                vec!["fingerprint-1".to_string()],
-                vec!["fingerprint-2".to_string(), "fingerprint-2b".to_string()],
-                vec!["fingerprint-3".to_string()],
-            ]
-        );
-    }
-
-    #[test]
-    fn kp_encrypted_shares_roster_rejects_wrong_share_ids() {
-        let err = KPEncryptedSharesRoster::new(vec![
-            test_kp_encrypted_shares(1),
-            test_kp_encrypted_shares(2),
-            test_kp_encrypted_shares(4),
-        ])
-        .expect_err("ids [1, 2, 4] are not exactly 1..=n");
-        assert!(format!("{err}").contains("share ids"), "{err}");
-    }
-
-    #[test]
-    fn kp_encrypted_shares_roster_replaces_one_recipient_and_preserves_the_rest() {
-        let shares = KPEncryptedSharesRoster::new(vec![
-            test_kp_encrypted_shares_for_fingerprints(1, vec!["a".into(), "b".into()]),
-            test_kp_encrypted_shares_for_fingerprints(2, vec!["c".into()]),
-        ])
-        .unwrap();
-
-        let (rotated, changed) = shares
-            .clone()
-            .replace_recipient("a", "d".into(), "new ciphertext".into())
-            .unwrap();
-        assert_eq!(changed.id.get(), 1);
-        assert_eq!(
-            changed
-                .ciphertexts_by_fingerprint
-                .keys()
-                .map(String::as_str)
-                .collect::<Vec<_>>(),
-            vec!["b", "d"]
-        );
-        assert_eq!(rotated.ciphertext_count(), 3);
-        assert!(rotated.find_by_fingerprint("a").is_none());
-        assert!(rotated.find_by_fingerprint("b").is_some());
-        assert!(rotated.find_by_fingerprint("c").is_some());
-        assert!(rotated.find_by_fingerprint("d").is_some());
-
-        let err = shares
-            .clone()
-            .replace_recipient("a", "c".into(), "collision".into())
-            .unwrap_err();
-        assert!(format!("{err}").contains("already present"), "{err}");
-
-        let err = shares
-            .replace_recipient("a", "a".into(), "same fingerprint".into())
-            .unwrap_err();
-        assert!(format!("{err}").contains("must differ"), "{err}");
-    }
-
-    #[test]
-    fn kp_encrypted_shares_roster_deserialize_through_validation() {
-        let json = serde_json::to_string(&vec![
-            test_kp_encrypted_shares(2),
-            test_kp_encrypted_shares(1),
-        ])
-        .unwrap();
-        let shares: KPEncryptedSharesRoster = serde_json::from_str(&json).unwrap();
-        let ids = shares.iter().map(|s| s.id.get()).collect::<Vec<_>>();
-        assert_eq!(ids, vec![1, 2]);
-
-        let bad_json = serde_json::to_string(&vec![
-            test_kp_encrypted_shares(1),
-            test_kp_encrypted_shares(1),
-        ])
-        .unwrap();
-        let err = serde_json::from_str::<KPEncryptedSharesRoster>(&bad_json).unwrap_err();
-        assert!(err.to_string().contains("encrypted share ids"), "{err}");
-    }
-
-    // Test eval function with specific coefficients
-    #[test]
-    fn test_eval_polynomial() {
-        // Test with simple polynomial: f(x) = 1 + 2x + 3x^2
-        let coefficients = vec![Scalar::ONE, Scalar::from(2u32), Scalar::from(3u32)];
-
-        // f(1) = 1 + 2(1) + 3(1)^2 = 6
-        let result1 = eval_poly(NonZeroU16::new(1).unwrap(), &coefficients);
-        assert_eq!(result1, Scalar::from(6u32));
-
-        // f(2) = 1 + 2(2) + 3(4) = 17
-        let result2 = eval_poly(NonZeroU16::new(2).unwrap(), &coefficients);
-        assert_eq!(result2, Scalar::from(17u32));
-    }
-}
