+++ crates/hashi-types/src/guardian/crypto/primitives/mod.rs
@@ -1,0 +1,8 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+mod hpke;
+mod secret_sharing;
+
+pub use hpke::*;
+pub use secret_sharing::*;
